audioGenerator/python_requirements.txt
pyyaml
numpy
scipy
python-rtmidi
PyFluidSynth
pydub
audioop-lts